Reorganize packages for further developments
