+++ digiforest_analysis/CMakeLists.txt
@@ -1,6 +1,6 @@
 cmake_minimum_required(VERSION 2.8.3)
 
-project(digiforest_drs)
+project(digiforest_analysis)
 add_compile_options(-std=c++14 -ggdb)
 
 find_package(catkin REQUIRED COMPONENTS
@@ -10,8 +10,8 @@ find_package(catkin REQUIRED COMPONENTS
 
 
 catkin_package(
-  INCLUDE_DIRS
-    include
+  # INCLUDE_DIRS
+  #   include
   CATKIN_DEPENDS 
     pcl_ros
     tf_conversions
@@ -21,7 +21,7 @@ catkin_package(
 ## Python #
 ###########
 catkin_python_setup()
-catkin_install_python(PROGRAMS src/digiforest_drs/generate_height_maps.py         
+catkin_install_python(PROGRAMS scripts/generate_height_maps.py         
                       DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION})
 
 ###########
@@ -29,15 +29,11 @@ catkin_install_python(PROGRAMS src/digiforest_drs/generate_height_maps.py
 ###########
 
 include_directories(
-  include
   ${catkin_INCLUDE_DIRS}
 )
 
 add_executable(generate_mesh src/tools/generate_mesh.cpp)
 target_link_libraries(generate_mesh ${catkin_LIBRARIES})
-
-add_executable(slam_logger_node src/slam_logger/slam_logger_node.cpp  src/slam_logger/slam_logger_tools.cpp)
-target_link_libraries(slam_logger_node ${catkin_LIBRARIES})
 
 
 #############
